[AIP-21] Move operators.pig_operator providers.apache.pig.operators.pig
