+++ airflow/providers/apache/pig/operators/__init__.py
@@ -16,28 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-"""Example DAG demonstrating the usage of the PigOperator."""
-
-from airflow.models import DAG
-from airflow.providers.apache.pig.operators.pig import PigOperator
-from airflow.utils.dates import days_ago
-
-args = {
-    'owner': 'airflow',
-    'start_date': days_ago(2),
-}
-
-dag = DAG(
-    dag_id='example_pig_operator',
-    default_args=args,
-    schedule_interval=None,
-    tags=['example']
-)
-
-run_this = PigOperator(
-    task_id="run_example_pig_script",
-    pig="ls /;",
-    pig_opts="-x local",
-    dag=dag,
-)
