+++ scripts/ci/pylint_todo.txt
@@ -155,7 +155,7 @@
 ./airflow/operators/mysql_operator.py
 ./airflow/operators/oracle_operator.py
 ./airflow/operators/papermill_operator.py
-./airflow/operators/pig_operator.py
+./airflow/providers/apache/pig/operators/pig.py
 ./airflow/operators/postgres_operator.py
 ./airflow/operators/presto_check_operator.py
 ./airflow/operators/presto_to_mysql.py
